refactor(113179): refatora header; refatora linha editais contratos
