+++ src/components/screens/Cadastros/EditaisContratosRefatorado/ConsultaEditaisContratos/index.tsx
@@ -1,28 +1,15 @@
 import { Spin } from "antd";
-import Botao from "components/Shareable/Botao";
-import {
-  BUTTON_ICON,
-  BUTTON_STYLE,
-  BUTTON_TYPE,
-} from "components/Shareable/Botao/constants";
-import { ToggleExpandir } from "components/Shareable/ToggleExpandir";
-import { deepCopy } from "helpers/utilities";
+
+import { Paginacao } from "components/Shareable/Paginacao";
 import HTTP_STATUS from "http-status-codes";
 import { ResponseEditalCotratoInterface } from "interfaces/responses.interface";
 import React, { useEffect, useState } from "react";
 import { Collapse } from "react-collapse";
+import { Form } from "react-final-form";
 import { getEditaisContratos } from "services/edital.service.js";
-import {
-  EditalContratoInterface,
-  EditalContratoListadoInterface,
-} from "../interfaces";
-import { Paginacao } from "components/Shareable/Paginacao";
+import { EditalContratoListadoInterface } from "../interfaces";
+import { Header, LinhaEditalContrato } from "./components";
 import "./style.scss";
-import { Tooltip } from "antd";
-import { VIGENCIA_STATUS } from "./constants";
-import InputText from "components/Shareable/Input/InputText";
-import { Field, Form } from "react-final-form";
-import { OnChange } from "react-final-form-listeners";
 
 export const ConsultaEditaisContratos = () => {
   const [editaisContratos, setEditaisContratos] =
@@ -32,10 +19,6 @@ export const ConsultaEditaisContratos = () => {
 
   const [loading, setLoading] = useState<boolean>(false);
   const [erro, setErro] = useState<string>("");
-
-  const PARAMS = { page };
-
-  let typingTimeout = null;
 
   const getEditaisContratosAsync = async (params = null): Promise<void> => {
     const response = await getEditaisContratos<ResponseEditalCotratoInterface>(
@@ -68,12 +51,6 @@ export const ConsultaEditaisContratos = () => {
     });
   };
 
-  const ativaContratoEdital = (index: number): void => {
-    const editaisContratos_ = deepCopy(editaisContratos);
-    editaisContratos_[index].ativo = !editaisContratos_[index].ativo;
-    setEditaisContratos(editaisContratos_);
-  };
-
   const nextPage = async ({
     page,
     values,
@@ -85,19 +62,6 @@ export const ConsultaEditaisContratos = () => {
     setLoading(true);
     await getEditaisContratosAsync({ page: page, buscar: values.buscar });
     setLoading(false);
-  };
-
-  const getStatusContrato = (
-    editalContrato: EditalContratoInterface
-  ): string => {
-    let status = "";
-    editalContrato.contratos.forEach((contrato) => {
-      const vigencia = contrato.vigencias[contrato.vigencias.length - 1];
-      if (vigencia.status === VIGENCIA_STATUS.VENCIDO) status = vigencia.status;
-      if (status !== "vencido" && vigencia.status === "proximo_ao_vencimento")
-        status = vigencia.status;
-    });
-    return status;
   };
 
   const onSubmit = () => {};
@@ -113,95 +77,22 @@ export const ConsultaEditaisContratos = () => {
             {({ handleSubmit, values }) => (
               <form onSubmit={handleSubmit}>
                 <div className="card">
-                  <div className={`row p-2 pt-3`}>
-                    <div className="col-3 title">Tipos de contratação</div>
-                    <div className="col-3 title">Nº do edital</div>
-                    <div className="col-3 title">
-                      Nº do processo administrativo
-                    </div>
-                    <div className="col-3 title">
-                      <Field
-                        component={InputText}
-                        name="buscar"
-                        placeholder="Pesquisar"
-                        className={`${BUTTON_STYLE.GRAY}`}
-                        icone={`${BUTTON_ICON.SEARCH} fa-lg`}
-                      />
-                      <OnChange name="buscar">
-                        {async (value) => {
-                          clearTimeout(typingTimeout);
-                          typingTimeout = setTimeout(async () => {
-                            setLoading(true);
-                            await getEditaisContratosAsync({
-                              busca: value,
-                              ...PARAMS,
-                            });
-                            setLoading(false);
-                            setPage(1);
-                          }, 1000);
-                        }}
-                      </OnChange>
-                    </div>
-                  </div>
+                  <Header
+                    getEditaisContratosAsync={getEditaisContratosAsync}
+                    setLoading={setLoading}
+                    page={page}
+                    setPage={setPage}
+                  />
                   <hr className="header" />
                   {editaisContratos.map((editalContrato, index) => {
                     return (
                       <div key={index} className="p-1">
-                        <div
-                          className={`row ${
-                            getStatusContrato(editalContrato) ===
-                            VIGENCIA_STATUS.PROXIMO_AO_VENCIMENTO
-                              ? VIGENCIA_STATUS.PROXIMO_AO_VENCIMENTO
-                              : ""
-                          }${
-                            getStatusContrato(editalContrato) ===
-                            VIGENCIA_STATUS.VENCIDO
-                              ? VIGENCIA_STATUS.VENCIDO
-                              : ""
-                          }`}
-                        >
-                          <div className="col-3">
-                            {editalContrato.tipo_contratacao}
-                          </div>
-                          <div className="col-3">{editalContrato.numero}</div>
-                          <div className="col-3">{editalContrato.processo}</div>
-                          <div className="col-3 d-flex my-auto">
-                            {getStatusContrato(editalContrato) ===
-                              VIGENCIA_STATUS.PROXIMO_AO_VENCIMENTO && (
-                              <Tooltip
-                                title={
-                                  "Data de vigência do contrato próxima ao vencimento, verifique se o contrato permanece ativo e adie a vigência"
-                                }
-                              >
-                                <div className="icon me-4 mt-1">
-                                  <span className="fas fa-exclamation" />
-                                </div>
-                              </Tooltip>
-                            )}
-                            {getStatusContrato(editalContrato) ===
-                              VIGENCIA_STATUS.VENCIDO && (
-                              <Tooltip
-                                title={
-                                  "Data de vigência do contrato expirada, verifique se o contrato permanece ativo e adie a vigência"
-                                }
-                              >
-                                <div className="icon orange me-4 mt-1">
-                                  <span className="fas fa-exclamation" />
-                                </div>
-                              </Tooltip>
-                            )}
-                            <ToggleExpandir
-                              onClick={() => ativaContratoEdital(index)}
-                              ativo={editalContrato.ativo}
-                              className="me-3"
-                            />
-                            <Botao
-                              type={BUTTON_TYPE.BUTTON}
-                              style={`${BUTTON_STYLE.GREEN_OUTLINE} no-border`}
-                              icon={BUTTON_ICON.EDIT}
-                            />
-                          </div>
-                        </div>
+                        <LinhaEditalContrato
+                          editalContrato={editalContrato}
+                          setEditaisContratos={setEditaisContratos}
+                          editaisContratos={editaisContratos}
+                          index={index}
+                        />
                         <hr />
                         <Collapse isOpened={editalContrato.ativo}>
                           <div className="row">
